+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file documents any relevant changes done to logics & vistache.
 This is the current development version.
 
 - Made Python3 compliant
+- Feature: Vistache extended to multiple `{{|}}`-blocks with different conditions
 - Feature: New function `lfill()` and `rfill()` implemented to fill strings with fill chars.
 
 ## [2.5.0] Vesuv